Update remove and clear all checked functionalities
