+++ webpack.config.js
@@ -8,7 +8,7 @@ module.exports = {
   },
   plugins: [
     new HtmlWebpackPlugin({
-        template: './src/index.html',
+      template: './src/index.html',
     }),
   ],
   output: {
@@ -23,39 +23,39 @@ module.exports = {
   mode: 'none',
   module: {
     rules: [
-        {
-          test: /\.css$/i,
-          use: ['style-loader', 'css-loader']
+      {
+        test: /\.css$/i,
+        use: ['style-loader', 'css-loader'],
+      },
+      {
+        test: /\.(png|svg|jpg|jpeg|gif)$/,
+        type: 'asset/resource',
+      },
+      {
+        test: /\.(woff|woff2|eot|ttf|otf)$/i,
+        type: 'asset/inline',
+      },
+      {
+        test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
+        include: path.resolve(__dirname, './node_modules/bootstrap-icons/font/fonts'),
+        use: {
+          loader: 'file-loader',
+          options: {
+            name: '[name].[ext]',
+            outputPath: 'webfonts',
+            publicPath: '../webfonts',
+          },
         },
-        {
-          test: /\.(png|svg|jpg|jpeg|gif)$/,
-          type: 'asset/resource',
-        },
-        {
-          test: /\.(woff|woff2|eot|ttf|otf)$/i,
-          type: "asset/inline",
-        },
-        {
-          test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-          include: path.resolve(__dirname, './node_modules/bootstrap-icons/font/fonts'),
-          use: {
-              loader: 'file-loader',
-              options: {
-                  name: '[name].[ext]',
-                  outputPath: 'webfonts',
-                  publicPath: '../webfonts',
-              },
-          }
       },
       {
         test: /\.woff($|\?)|\.woff2($|\?)|\.ttf($|\?)|\.eot($|\?)|\.svg($|\?)/i,
         type: 'asset/resource',
         generator: {
-            //filename: 'fonts/[name]-[hash][ext][query]'
-            filename: 'fonts/[name][ext][query]'
-        }
-    },
+          // filename: 'fonts/[name]-[hash][ext][query]'
+          filename: 'fonts/[name][ext][query]',
+        },
+      },
     ],
   },
-  
+
 };
